% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{apply_facets}
\alias{apply_facets}
\title{Apply facets to to \code{ggplot2::ggplot}}
\usage{
apply_facets(
  p,
  formula,
  facets = TRUE,
  nrow = NULL,
  ncol = 1,
  scales = "free_y",
  ...
)
}
\arguments{
\item{p}{\code{ggplot2::ggplot} instance}

\item{formula}{\code{stats::formula} instance}

\item{facets}{Logical value to specify use facets}

\item{nrow}{Number of facet/subplot rows}

\item{ncol}{Number of facet/subplot columns}

\item{scales}{Scale value passed to \code{ggplot2}}

\item{...}{other arguments passed to methods}
}
\value{
ggplot
}
\description{
Apply facets to to \code{ggplot2::ggplot}
}
