% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{apply_grid}
\alias{apply_grid}
\title{Apply grid to to \code{ggplot2::ggplot}}
\usage{
apply_grid(p, formula, scales = "free_y", ...)
}
\arguments{
\item{p}{\code{ggplot2::ggplot} instance}

\item{formula}{\code{stats::formula} instance}

\item{scales}{Scale value passed to \code{ggplot2}}

\item{...}{other arguments passed to methods}
}
\description{
Apply grid to to \code{ggplot2::ggplot}
}
