% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.density}
\alias{autoplot.density}
\title{Plot \code{stats::density}}
\usage{
\method{autoplot}{density}(data, p = NULL, colour = "#000000",
  linetype = "solid", fill = NULL, alpha = 0.3, ...)
}
\arguments{
\item{data}{\code{stats::density} instance}

\item{p}{\code{ggplot2::ggplot} instance to plot}

\item{colour}{Line colour}

\item{linetype}{Line type}

\item{fill}{Fill colour}

\item{alpha}{Alpha}

\item{...}{Keywords passed to PDC/CDF func}
}
\value{
ggplot
}
\description{
Plot \code{stats::density}
}
\examples{
ggplot2::autoplot(stats::density(stats::rnorm(1:50)))
}

