% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_forecast.R
\name{autoplot.ets}
\alias{autoplot.bats}
\alias{autoplot.ets}
\title{Autoplot \code{forecast::bats} and \code{forecast::ets}}
\usage{
\method{autoplot}{ets}(object, columns = NULL, ...)
}
\arguments{
\item{object}{\code{forecast::bats} and \code{forecast::ets}  instance}

\item{columns}{Character vector specifies target column name(s)}

\item{...}{other arguments passed to \code{autoplot.ts}}
}
\value{
ggplot
}
\description{
Autoplot \code{forecast::bats} and \code{forecast::ets}
}
\examples{
d.bats <- forecast::bats(UKgas)
ggplot2::autoplot(d.bats)
ggplot2::autoplot(d.bats, columns = 'Residuals')
ggplot2::autoplot(forecast::ets(UKgas))
}

