% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{autoplot.ggmultiplot}
\alias{autoplot.ggmultiplot}
\title{Autoplot \code{ggmultiplot} instances.
It returns the passed instance as it is.}
\usage{
\method{autoplot}{ggmultiplot}(object, ...)
}
\arguments{
\item{object}{ggmultiplot instance}

\item{...}{Not used.}
}
\value{
ggmultiplot
}
\description{
Autoplot \code{ggmultiplot} instances.
It returns the passed instance as it is.
}
