% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{autoplot.ggplot}
\alias{autoplot.ggplot}
\title{Autoplot \code{ggplot} instances.
It returns the passed instance as it is.}
\usage{
\method{autoplot}{ggplot}(object, ...)
}
\arguments{
\item{object}{ggplot instance}

\item{...}{Not used.}
}
\value{
ggplot
}
\description{
Autoplot \code{ggplot} instances.
It returns the passed instance as it is.
}
