% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_stats_tests.R
\name{autoplot.htest}
\alias{autoplot.htest}
\title{Autoplot \code{stats::htest}.}
\usage{
\method{autoplot}{htest}(data)
}
\arguments{
\item{data}{\code{stats::htest} instance}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::htest}.
}
\examples{
ggplot2::autoplot(t.test(rnorm(10)))
ggplot2::autoplot(t.test(mpg ~ am, data = mtcars))
ggplot2::autoplot(wilcox.test(mpg ~ am, data = mtcars, conf.int = TRUE))
ggplot2::autoplot(cor.test(mtcars$wt, mtcars$mpg))
}

