% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.kmeans}
\alias{autoplot.kmeans}
\title{Autoplot \code{stats::kmeans}.}
\usage{
\method{autoplot}{kmeans}(data, original = NULL)
}
\arguments{
\item{data}{\code{stats::princomp} instance}

\item{original}{Original data used for K-means. Mandatory for plotting.}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::kmeans}.
}
\examples{
df <- iris[c(1, 2, 3, 4)]
ggplot2::autoplot(stats::princomp(df), original = iris)
}

