% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base_fortify_list.R
\name{autoplot.list}
\alias{autoplot.list}
\title{Autoplot list.}
\usage{
\method{autoplot}{list}(object, ...)
}
\arguments{
\item{object}{\code{list} instance}

\item{...}{other arguments passed to methods}
}
\value{
ggplot
}
\description{
Autoplot list.
}

