% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_maps.R
\name{autoplot.map}
\alias{autoplot.map}
\title{Autoplot \code{maps::map}}
\usage{
\method{autoplot}{map}(
  object,
  p = NULL,
  geom = "path",
  group = "group",
  colour = "black",
  size = NULL,
  linetype = NULL,
  alpha = NULL,
  fill = NULL,
  shape = NULL,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  log = "",
  main = NULL,
  xlab = "",
  ylab = "",
  asp = NULL,
  ...
)
}
\arguments{
\item{object}{\code{maps::map} instance}

\item{p}{\code{ggplot2::ggplot} instance}

\item{geom}{geometric string for map. 'path', 'point' or 'polygon'}

\item{group}{key for grouping geoms}

\item{colour}{line colour}

\item{size}{point size}

\item{linetype}{line type}

\item{alpha}{alpha}

\item{fill}{fill colour}

\item{shape}{point shape}

\item{xlim}{limits for x axis}

\item{ylim}{limits for y axis}

\item{log}{which variables to log transform ("x", "y", or "xy")}

\item{main}{character vector or expression for plot title}

\item{xlab}{character vector or expression for x axis label}

\item{ylab}{character vector or expression for y axis label}

\item{asp}{the y/x aspect ratio}

\item{...}{other arguments passed to methods}
}
\value{
ggplot
}
\description{
Autoplot \code{maps::map}
}
