% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_performance.R
\name{autoplot.performance}
\alias{autoplot.performance}
\title{Autoplot \code{ROCR::performance}}
\usage{
\method{autoplot}{performance}(object, p = NULL, ...)
}
\arguments{
\item{object}{\code{ROCR::performance} instance}

\item{p}{\code{ggplot2::ggplot} instances}

\item{...}{other arguments passed to methods}
}
\value{
ggplot
}
\description{
Autoplot \code{ROCR::performance}
}

