% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_raster.R
\name{autoplot.RasterCommon}
\alias{autoplot.RasterCommon}
\title{Autoplot \code{raster::raster}}
\usage{
\method{autoplot}{RasterCommon}(
  object,
  raster.layer = NULL,
  p = NULL,
  alpha = NULL,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  log = "",
  main = NULL,
  xlab = "",
  ylab = "",
  asp = NULL,
  ...
)
}
\arguments{
\item{object}{\code{raster::raster} instance}

\item{raster.layer}{name of the layer to plot}

\item{p}{\code{ggplot2::ggplot} instance}

\item{alpha}{alpha}

\item{xlim}{limits for x axis}

\item{ylim}{limits for y axis}

\item{log}{which variables to log transform ("x", "y", or "xy")}

\item{main}{character vector or expression for plot title}

\item{xlab}{character vector or expression for x axis label}

\item{ylab}{character vector or expression for y axis label}

\item{asp}{the y/x aspect ratio}

\item{...}{other arguments passed to methods}
}
\value{
ggplot
}
\description{
Only plot the first layer of the given raster
}
