% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_cluster.R
\name{autoplot.silhouette}
\alias{autoplot.silhouette}
\title{Autoplot silhouette instances}
\usage{
\method{autoplot}{silhouette}(object, ...)
}
\arguments{
\item{object}{Silhouette instance}

\item{...}{other arguments passed to \code{ggplot2::ggplot}}
}
\value{
ggplot
}
\description{
Autoplot silhouette instances
}
\examples{
model = cluster::pam(iris[-5], 3L)
sil = cluster::silhouette(model)
autoplot(sil)
}
