% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.stl}
\alias{autoplot.stl}
\title{Autoplot \code{stats::stl}.}
\usage{
\method{autoplot}{stl}(data)
}
\arguments{
\item{data}{\code{stats::stl} instance}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::stl}.
}
\examples{
d.stl <- stats::stl(AirPassengers, s.window = 'periodic')
ggplot2::autoplot(d.stl)
}

