% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base_fortify_ts.R
\name{autoplot.ts}
\alias{autoplot.decomposed.ts}
\alias{autoplot.irts}
\alias{autoplot.stl}
\alias{autoplot.timeSeries}
\alias{autoplot.ts}
\alias{autoplot.xts}
\title{Autoplot time-series-like.}
\usage{
\method{autoplot}{ts}(object, columns = NULL, group = NULL,
  is.date = NULL, index.name = "Index", p = NULL, ts.scale = FALSE,
  scales = "free_y", facet = TRUE, facets = facet, nrow = NULL,
  ncol = 1, ts.geom = "line", ts.colour = "#000000",
  ts.linetype = "solid", xlab = "", ylab = "", ...)
}
\arguments{
\item{object}{time-series-like instance}

\item{columns}{Character vector specifies target column name(s)}

\item{group}{Character vector specifies grouping}

\item{is.date}{Logical frag indicates whether the \code{stats::ts} is date or not.
If not provided, regard the input as date when the frequency is 4 or 12.}

\item{index.name}{Specify column name for time series index when passing \code{data.frame} via data.}

\item{p}{\code{ggplot2::ggplot} instance}

\item{ts.scale}{Logical flag indicating whether to perform scaling each timeseries}

\item{scales}{Scale value passed to \code{ggplot2}}

\item{facet}{(Deprecated) use facets}

\item{facets}{Logical value to specify use facets for multivariate time-series}

\item{nrow}{Number of facet/subplot rows}

\item{ncol}{Number of facet/subplot columns}

\item{ts.geom}{Geometric string for time-series plotting. 'line', 'bar' or 'point'.}

\item{ts.colour}{Line colour for time-series}

\item{ts.linetype}{Line type for time-series}

\item{xlab}{Character vector or expression for x axis label}

\item{ylab}{Character vector or expression for y axis label}

\item{...}{other arguments passed to methods}
}
\value{
ggplot
}
\description{
Autoplot time-series-like.
}
\examples{
data(Canada, package = 'vars')
autoplot(AirPassengers)
autoplot(UKgas, ts.geom = 'bar')
autoplot(Canada)
autoplot(Canada, columns = 'e', is.date = TRUE)
autoplot(Canada, facets = FALSE)

library(zoo)
autoplot(xts::as.xts(AirPassengers))
autoplot(xts::as.xts(UKgas))
autoplot(xts::as.xts(Canada))

autoplot(timeSeries::as.timeSeries(AirPassengers))
autoplot(timeSeries::as.timeSeries(Canada))

its <- tseries::irts(cumsum(rexp(10, rate = 0.1)), matrix(rnorm(20), ncol=2))
autoplot(its)

autoplot(stats::stl(UKgas, s.window = 'periodic'))
autoplot(stats::decompose(UKgas))
}

