% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.varprd}
\alias{autoplot.varprd}
\title{Autoplot \code{vars::varprd}.}
\usage{
\method{autoplot}{varprd}(data)
}
\arguments{
\item{data}{\code{vars::varpred} instance}
}
\value{
ggplot
}
\description{
Autoplot \code{vars::varprd}.
}
\examples{
data(Canada, package = 'vars')
d.vselect <- vars::VARselect(Canada, lag.max = 5, type = 'const')$selection[1]
d.var <- vars::VAR(Canada, p = d.vselect, type = 'const')
ggplot2::autoplot(stats::predict(d.var, n.ahead = 50))
}

