% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cbind.original}
\alias{cbind.original}
\title{Wrapper for cbind}
\usage{
cbind.original(data, original = NULL)
}
\arguments{
\item{data}{\code{data.frame} instance}

\item{original}{Joined to data if provided.
cluster labels to the original}
}
\value{
list
}
\description{
Wrapper for cbind
}
\examples{
ggfortify:::cbind.original(iris[1:2], iris[3:5])
}

