% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{cbind_wraps}
\alias{cbind_wraps}
\title{Wrapper for cbind}
\usage{
cbind_wraps(df1, df2)
}
\arguments{
\item{data}{\code{data.frame} instance}

\item{original}{Joined to data if provided.
cluster labels to the original}
}
\value{
list
}
\description{
Wrapper for cbind
}
\examples{
ggfortify:::cbind_wraps(iris[1:2], iris[3:5])
}

