% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base_fortify_list.R
\name{check.attrs}
\alias{check.attrs}
\title{Infer class name}
\usage{
check.attrs(data, expected)
}
\arguments{
\item{data}{list instance}

\item{expected}{expected character vector}
}
\value{
logical
}
\description{
Infer class name
}

