% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_fortify_list.R
\name{check_names}
\alias{check_names}
\title{Check data names are equal with expected}
\usage{
check_names(data, expected)
}
\arguments{
\item{data}{\code{list} instance to be checked}

\item{expected}{expected character vector}
}
\value{
logical
}
\description{
Check data names are equal with expected
}
