% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslib.R
\name{confint.acf}
\alias{confint.acf}
\title{Calcurate confidence interval for \code{stats::acf}}
\usage{
\method{confint}{acf}(x, ci = 0.95, ci.type = "white")
}
\arguments{
\item{x}{\code{stats::acf} instance}

\item{ci}{Float value for confidence interval}

\item{ci.type}{"white" or "ma"}
}
\value{
vector
}
\description{
Calcurate confidence interval for \code{stats::acf}
}
\examples{
air.acf <- acf(AirPassengers, plot = FALSE)
ggfortify:::confint.acf(air.acf)
ggfortify:::confint.acf(air.acf, ci.type = 'ma')
}

