% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fitted.ar}
\alias{fitted.ar}
\title{Calcurate fitted values for \code{stats::ar}}
\usage{
\method{fitted}{ar}(object, ...)
}
\arguments{
\item{object}{\code{stats::ar} instance}
}
\value{
ts An time series of the one-step forecasts
}
\description{
Calcurate fitted values for \code{stats::ar}
}
\examples{
fitted(ar(WWWusage))
}

