% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fortify_surv.R
\name{fortify.aareg}
\alias{fortify.aareg}
\title{Convert \code{survival::aareg} to \code{data.frame}}
\usage{
\method{fortify}{aareg}(x, maxtime, surv.connect = TRUE, melt = FALSE)
}
\arguments{
\item{maxtime}{}

\item{surv.connect}{logical frag indicates whether connects survival curve to the origin}

\item{model}{\code{survival::aareg} instance}
}
\value{
data.frame
}
\description{
Convert \code{survival::aareg} to \code{data.frame}
}
\examples{
library(survival)
fortify(aareg(Surv(time, status) ~ age + sex + ph.ecog, data = lung, nmin = 1))
fortify(aareg(Surv(time, status) ~ age + sex + ph.ecog, data = lung, nmin = 1), melt = TRUE)
}

