% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.Arima}
\alias{fortify.Arima}
\alias{fortify.ar}
\title{Convert \code{stats::arima} to data.frame.}
\usage{
\method{fortify}{Arima}(data)
}
\arguments{
\item{data}{\code{stats::arima} instance}
}
\value{
data.frame
}
\description{
Convert \code{stats::arima} to data.frame.
}
\examples{
ggplot2::fortify(stats::arima(UKgas))
ggplot2::fortify(stats::ar(AirPassengers))
}

