% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_glmnet.R
\name{fortify.cv.glmnet}
\alias{fortify.cv.glmnet}
\title{Convert \code{glmnet::cv.glmnet} to \code{data.frame}}
\usage{
\method{fortify}{cv.glmnet}(model, data = NULL, ...)
}
\arguments{
\item{model}{\code{glmnet::cv.glmnet} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{glmnet::cv.glmnet} to \code{data.frame}
}
\examples{
fortify(glmnet::cv.glmnet(data.matrix(Orange[-3]), data.matrix(Orange[3])))
}
