% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.density}
\alias{fortify.density}
\title{Convert \code{stats::density} to \code{data.frame}.}
\usage{
\method{fortify}{density}(data)
}
\arguments{
\item{data}{\code{stats::density} instance}
}
\value{
data.frame
}
\description{
Convert \code{stats::density} to \code{data.frame}.
}
\examples{
ggplot2::fortify(stats::density(stats::rnorm(1:50)))
}

