% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.dist}
\alias{fortify.dist}
\title{Convert \code{stats::dist} to data.frame.}
\usage{
\method{fortify}{dist}(data, original = NULL)
}
\arguments{
\item{data}{\code{stats::dist} instance}
}
\value{
data.frame
}
\description{
Convert \code{stats::dist} to data.frame.
}
\examples{
ggplot2::fortify(eurodist)
}

