% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_forecast.R
\name{fortify.ets}
\alias{fortify.ets}
\title{Convert \code{forecast::bats} and \code{forecast::ets} to data.frame.}
\usage{
\method{fortify}{ets}(data)
}
\arguments{
\item{data}{\code{forecast::bats} or \code{forecast::ets} instance}
}
\value{
data.frame
}
\description{
Convert \code{forecast::bats} and \code{forecast::ets} to data.frame.
}
\examples{
ggplot2::fortify(forecast::bats(UKgas))
ggplot2::fortify(forecast::ets(UKgas))
}

