% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.forecast}
\alias{fortify.forecast}
\title{Convert \code{forecast::forecast} to data.frame.}
\usage{
\method{fortify}{forecast}(data)
}
\arguments{
\item{data}{\code{forecast::forecast} instance}
}
\value{
data.frame
}
\description{
Convert \code{forecast::forecast} to data.frame.
}
\examples{
d.arima <- forecast::auto.arima(AirPassengers)
d.forecast <- forecast::forecast(d.arima, level = c(95), h = 50)
ggplot2::fortify(d.forecast)
}

