% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_stats_tests.R
\name{fortify.htest}
\alias{fortify.htest}
\title{Convert \code{stats::htest} to data.frame.}
\usage{
\method{fortify}{htest}(data)
}
\arguments{
\item{data}{\code{stats::htest} instance}
}
\value{
data.frame
}
\description{
Convert \code{stats::htest} to data.frame.
}
\examples{
ggplot2::fortify(t.test(rnorm(10)))
ggplot2::fortify(t.test(mpg ~ am, data = mtcars))
ggplot2::fortify(wilcox.test(mpg ~ am, data = mtcars))
ggplot2::fortify(cor.test(mtcars$wt, mtcars$mpg))
}

