% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fortify_stats.R
\name{fortify.lfda}
\alias{fortify.lfda}
\alias{fortify.princomp}
\title{Convert \code{lfda::lfda} or \code{lfda::klfda} or \code{lfda::self} to \code{data.frame}}
\usage{
\method{fortify}{lfda}(model, data = NULL, original = NULL, ...)
}
\arguments{
\item{model}{\code{lfda::lfda} or \code{lfda::klfda} or \code{lfda::self} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{lfda::lfda} or \code{lfda::klfda} or \code{lfda::self} to \code{data.frame}
}
\examples{
k <- iris[,-5]
y <- iris[,5]
r <- 3
model <- lfda::lfda(k,y,r,metric="plain")
fortify(model)
}

