% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base_fortify_list.R
\name{fortify.list}
\alias{fortify.list}
\title{Convert list to data.frame.}
\usage{
\method{fortify}{list}(data)
}
\arguments{
\item{data}{list instance}
}
\value{
data.frame
}
\description{
Convert list to data.frame.
}

