% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.matrix}
\alias{fortify.matrix}
\title{Convert \code{base::matrix} to data.frame.}
\usage{
\method{fortify}{matrix}(data)
}
\arguments{
\item{data}{\code{base::matrix} instance}
}
\value{
data.frame
}
\description{
Convert \code{base::matrix} to data.frame.
}
\examples{
ggplot2::fortify(matrix(1:6, nrow=2, ncol=3))
}

