% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_performance.R
\name{fortify.performance}
\alias{fortify.performance}
\title{Convert \code{ROCR::performance} objects to \code{data.frame}}
\usage{
\method{fortify}{performance}(model, data = NULL, ...)
}
\arguments{
\item{model}{\code{performance} instances}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{ROCR::performance} objects to \code{data.frame}
}
