% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.prcomp}
\alias{fortify.prcomp}
\title{Convert \code{stats::prcomp} to data.frame.}
\usage{
\method{fortify}{prcomp}(data, original = NULL)
}
\arguments{
\item{data}{\code{stats::prcomp} instance}

\item{original}{Joined to PCA result if provided. Intended to be used for attaching
non-numeric values original data has. Numeric values are automatically attached.}
}
\value{
data.frame
}
\description{
Convert \code{stats::prcomp} to data.frame.
}
\examples{
df <- iris[c(1, 2, 3, 4)]
ggplot2::fortify(stats::prcomp(df))
ggplot2::fortify(stats::prcomp(df), original = iris)
}

