% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.princomp}
\alias{fortify.princomp}
\title{Convert \code{stats::princomp} to data.frame.}
\usage{
\method{fortify}{princomp}(data)
}
\arguments{
\item{data}{\code{stats::princomp} instance}
}
\value{
data.frame
}
\description{
Convert \code{stats::princomp} to data.frame.
}
\examples{
df <- iris[c(1, 2, 3, 4)]
ggplot2::fortify(stats::princomp(df))
}

