% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_raster.R
\name{fortify.RasterCommon}
\alias{fortify.RasterCommon}
\title{Convert \code{raster} to \code{data.frame}}
\usage{
\method{fortify}{RasterCommon}(model, data = NULL, maxpixels = 1e+05, rename = TRUE, ...)
}
\arguments{
\item{model}{\code{raster} instances}

\item{data}{original dataset, if needed}

\item{maxpixels}{number of pixels for resampling}

\item{rename}{logical flag indicating whether to rename coordinates to long and lat}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{raster} to \code{data.frame}
}
