% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_stats.R
\name{fortify.spec}
\alias{fortify.spec}
\title{Convert \code{stats::spec} to data.frame.}
\usage{
\method{fortify}{spec}(model, data, ...)
}
\arguments{
\item{model}{\code{stats::spec} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{stats::spec} to data.frame.
}
\examples{
ggplot2::fortify(spectrum(AirPassengers))
ggplot2::fortify(stats::spec.ar(AirPassengers))
ggplot2::fortify(stats::spec.pgram(AirPassengers))
}

