% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_stats.R
\name{fortify.stepfun}
\alias{fortify.stepfun}
\title{Convert \code{stats::stepfun} to \code{data.frame}}
\usage{
\method{fortify}{stepfun}(model, data, ...)
}
\arguments{
\item{model}{\code{stats::stepfun} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{stats::stepfun} to \code{data.frame}
}
\examples{
fortify(stepfun(c(1, 2, 3), c(4, 5, 6, 7)))
fortify(stepfun(c(1), c(4, 5)))
fortify(stepfun(c(1, 3, 4, 8), c(4, 5, 2, 3, 5)))
fortify(stepfun(c(1, 2, 3, 4, 5, 6, 7, 8, 10), c(4, 5, 6, 7, 8, 9, 10, 11, 12, 9)))
}
