% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.stl}
\alias{fortify.HoltWinters}
\alias{fortify.decomposed.ts}
\alias{fortify.stl}
\title{Convert \code{stats::stl}, \code{stats::decomposed.ts} and
 \code{stats::HoltWinters} to data.frame.}
\usage{
\method{fortify}{stl}(data)
}
\arguments{
\item{data}{\code{stats::stl}, \code{stats::decomposed.ts} or \code{stats::HoltWinters} instance}
}
\value{
data.frame
}
\description{
Convert \code{stats::stl}, \code{stats::decomposed.ts} and
 \code{stats::HoltWinters} to data.frame.
}
\examples{
ggplot2::fortify(stats::stl(UKgas, s.window = 'periodic'))
ggplot2::fortify(stats::decompose(UKgas))
ggplot2::fortify(stats::HoltWinters(USAccDeaths))
}

