% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.ts}
\alias{fortify.ts}
\title{Convert time-series-like to data.frame.}
\usage{
\method{fortify}{ts}(data, index.name = "Index", data.name = "Data",
  is.date = NULL, columns = NULL, melt = FALSE)
}
\arguments{
\item{data}{\code{stats::ts}, \code{timeSeries::timeSeries} or \code{tseries::irts} instance}

\item{index.name}{Specify column name for time series index}

\item{is.date}{Logical frag indicates whether the \code{stats::ts} is date or not.
If not provided, regard the input as date when the frequency is 4 or 12.}

\item{melt}{Logical flag indicating whether to melt each timeseries as variable}
}
\value{
data.frame
}
\description{
Convert time-series-like to data.frame.
}
\examples{
ggplot2::fortify(AirPassengers)
ggplot2::fortify(timeSeries::as.timeSeries(AirPassengers))

its <- tseries::irts(cumsum(rexp(10, rate = 0.1)), matrix(rnorm(20), ncol=2))
ggplot2::fortify(its)
}

