% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.tsmodel}
\alias{fortify.Arima}
\alias{fortify.HoltWinters}
\alias{fortify.ar}
\alias{fortify.fGARCH}
\alias{fortify.fracdiff}
\alias{fortify.nnetar}
\alias{fortify.tsmodel}
\title{Convert time series models (like AR, ARIMA) to data.frame.}
\usage{
\method{fortify}{tsmodel}(data, original = NULL, is.date = NULL)
}
\arguments{
\item{data}{\code{stats::ar}, \code{stats::Arima}, \code{stats::HoltWinters},
\code{fracdiff::fracdiff}, \code{forecast::nnetar} or \code{fGarch::fGARCH}instance}

\item{original}{Original data for \code{stats::ar}, \code{stats::Arima}.
Not used in other models.
If not provided, try to retrieve from current environment using variable name.}

\item{is.date}{Logical frag indicates whether the \code{stats::ts} is date or not.
If not provided, regard the input as date when the frequency is 4 or 12.}
}
\value{
data.frame
}
\description{
Convert time series models (like AR, ARIMA) to data.frame.
}
\examples{
ggplot2::fortify(stats::ar(AirPassengers))
ggplot2::fortify(stats::arima(UKgas))
ggplot2::fortify(stats::arima(UKgas), original = UKgas, is.date = TRUE)
ggplot2::fortify(forecast::auto.arima(austres))
ggplot2::fortify(forecast::arfima(AirPassengers))
ggplot2::fortify(forecast::nnetar(UKgas))
ggplot2::fortify(stats::HoltWinters(USAccDeaths))

data(LPP2005REC, package = 'timeSeries')
x = timeSeries::as.timeSeries(LPP2005REC)
d.Garch = fGarch::garchFit(LPP40 ~ garch(1, 1), data = 100 * x, trace = FALSE)
ggplot2::fortify(d.Garch)
}

