% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_base.R
\name{fortify_base}
\alias{fortify_base}
\title{Convert \code{base::table} to \code{data.frame}}
\usage{
fortify_base(model, data, ...)
}
\arguments{
\item{model}{\code{base::table} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{base::table} to \code{data.frame}
}
