% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_maps.R
\name{fortify_map}
\alias{fortify_map}
\title{Convert \code{maps::map} to \code{data.frame}.}
\usage{
fortify_map(model, data = NULL, ...)
}
\arguments{
\item{model}{\code{maps::map} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{maps::map} to \code{data.frame}.
}
