% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\docType{data}
\name{geom_confint}
\alias{GeomConfint}
\alias{geom_confint}
\title{Connect observations by stairs.}
\format{An object of class \code{GeomConfint} (inherits from \code{GeomRibbon}, \code{Geom}, \code{ggproto}) of length 4.}
\usage{
geom_confint(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, ...)

GeomConfint
}
\arguments{
\item{mapping}{the aesthetic mapping}

\item{data}{a layer specific dataset}

\item{stat}{the statistical transformation to use on the data for this layer}

\item{position}{the position adjustment to use for overlapping points on this layer}

\item{na.rm}{logical frag whether silently remove missing values}

\item{...}{other arguments passed to methods}
}
\description{
Connect observations by stairs.
}
\keyword{datasets}

