% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotlib.R
\name{geom_factory}
\alias{geom_factory}
\title{Factory function to control \code{ggplot2::geom_xxx} functions}
\usage{
geom_factory(geomfunc, data, ...)
}
\arguments{
\item{geomfunc}{\code{ggplot2::geom_xxx} function}

\item{data}{plotting data}

\item{...}{other arguments passed to methods}
}
\value{
proto
}
\description{
Factory function to control \code{ggplot2::geom_xxx} functions
}

