% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.dtindex.continuous}
\alias{get.dtindex.continuous}
\title{Get \code{Date} \code{vector} continue to \code{ts} index.}
\usage{
get.dtindex.continuous(data, length)
}
\arguments{
\item{data}{\code{ts} instance}

\item{length}{A number to continue}
}
\value{
vector
}
\description{
Get \code{Date} \code{vector} continue to \code{ts} index.
}
\examples{
ggfortify:::get.dtindex.continuous(AirPassengers, length = 10)
ggfortify:::get.dtindex.continuous(UKgas, length = 10)
}

