% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.dtindex}
\alias{get.dtindex}
\title{Convert \code{ts} index to \code{Date} \code{vector}.}
\usage{
get.dtindex(data)
}
\arguments{
\item{data}{\code{ts} instance}
}
\value{
vector
}
\description{
Convert \code{ts} index to \code{Date} \code{vector}.
}
\examples{
ggfortify:::get.dtindex(AirPassengers)
ggfortify:::get.dtindex(UKgas)
}

