% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{get.layout}
\alias{get.layout}
\title{Calcurate layout matrix for \code{ggmultiplot}}
\usage{
get.layout(nplots, ncol, nrow)
}
\arguments{
\item{nplots}{Number of plots}

\item{ncol}{Number of grid columns}

\item{nrow}{Number of grid rows}
}
\value{
matrix
}
\description{
Calcurate layout matrix for \code{ggmultiplot}
}
\examples{
ggfortify:::get.layout(3, 2, 2)
}
