% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{get_geom_function}
\alias{get_geom_function}
\title{Factory function to control \code{ggplot2::geom_xxx} functions}
\usage{
get_geom_function(geom, allowed = c("line", "bar", "point"))
}
\arguments{
\item{geom}{string representation of \code{ggplot2::geom_xxx} function}

\item{allowed}{character vector contains allowed values}
}
\value{
function
}
\description{
Factory function to control \code{ggplot2::geom_xxx} functions
}
\examples{
ggfortify:::get_geom_function('point')
ggfortify:::get_geom_function('line', allowed = c('line'))
}
