% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggcpgram}
\alias{ggcpgram}
\title{Plots a cumulative periodogram.}
\usage{
ggcpgram(ts, taper = 0.1, colour = "#000000", linetype = "solid",
  conf.int = TRUE, conf.int.colour = "#0000FF",
  conf.int.linetype = "dashed", conf.int.fill = NULL,
  conf.int.alpha = 0.3)
}
\arguments{
\item{taper}{Proportion tapered in forming the periodogram}

\item{colour}{Line colour}

\item{linetype}{Line type}

\item{conf.int}{Logical flag indicating whether to plot confidence intervals}

\item{conf.int.colour}{Line colour for confidence intervals}

\item{conf.int.linetype}{Line type for confidence intervals}

\item{conf.int.fill}{Fill colour for confidence intervals}

\item{conf.int.alpha}{Alpha for confidence intervals}

\item{x}{\code{stats::acf} instance}
}
\value{
ggplot
}
\description{
Plots a cumulative periodogram.
}
\examples{
ggcpgram(AirPassengers)
}

