% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfortify.R
\docType{package}
\name{ggfortify}
\alias{ggfortify-package}
\alias{ggfortify}
\title{ggfortify}
\description{
Define Fortify and Autoplot to Allow 'ggplot2' to Draw Some Popular Packages
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sinhrks/ggfortify}
  \item Report bugs at \url{https://github.com/sinhrks/ggfortify/issues}
}

}
\author{
\strong{Maintainer}: Yuan Tang \email{terrytangyuan@gmail.com} (\href{https://orcid.org/0000-0001-5243-233X}{ORCID})

Authors:
\itemize{
  \item Masaaki Horikoshi \email{sinhrks@gmail.com}
}

Other contributors:
\itemize{
  \item Austin Dickey [contributor]
  \item Matthias Grenié [contributor]
  \item Ryan Thompson [contributor]
  \item Luciano Selzer [contributor]
  \item Dario Strbenac [contributor]
  \item Kirill Voronin [contributor]
  \item Damir Pulatov [contributor]
}

}
