% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tslib.R
\name{gglagplot}
\alias{gglagplot}
\title{Plot time series against lagged versions of themselves.}
\usage{
gglagplot(ts, lags = 1, nrow = NULL, ncol = NULL)
}
\arguments{
\item{ts}{\code{stats::ts} instance}

\item{lags}{Number of lag plots desired}

\item{nrow}{Number of plot rows}

\item{ncol}{Number of plot columns}
}
\value{
ggplot
}
\description{
Plot time series against lagged versions of themselves.
}
\examples{
gglagplot(AirPassengers)
}

