% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotlib.R
\docType{class}
\name{ggmultiplot-class}
\alias{ggmultiplot-class}
\title{An S4 class to hold multiple \code{ggplot2::ggplot} instances.}
\description{
An S4 class to hold multiple \code{ggplot2::ggplot} instances.
}
\section{Slots}{

\describe{
\item{\code{plots}}{List of \code{ggplot2::ggplot} instances}

\item{\code{ncol}}{Number of grid columns}

\item{\code{nrow}}{Number of grid rows}
}}

