% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{grid.draw.ggmultiplot}
\alias{grid.draw.ggmultiplot}
\title{The implemented grid.draw method for ggmultiplot, in order to work
with ggsave() properly}
\usage{
\method{grid.draw}{ggmultiplot}(plot)
}
\arguments{
\item{plot}{\code{ggmultiplot}}
}
\description{
The implemented grid.draw method for ggmultiplot, in order to work
with ggsave() properly
}
