% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_derived_from}
\alias{is_derived_from}
\title{Check object is target class, or object is \code{data.frame} fortified from target.}
\usage{
is_derived_from(object, target)
}
\arguments{
\item{object}{instance to be checked. For data.frame, check whether it is fortified from target class}

\item{target}{class name}
}
\value{
logical
}
\description{
Check object is target class, or object is \code{data.frame} fortified from target.
}
\examples{
ggfortify:::is_derived_from(prcomp(iris[-5]), 'prcomp')
}
