% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse.formula}
\alias{parse.formula}
\title{Expand \code{stats::formula} expression}
\usage{
parse.formula(formula)
}
\arguments{
\item{data}{Scaled data}

\item{center}{Centered vector}

\item{scale}{Scale vector}
}
\value{
data.frame
}
\description{
Expand \code{stats::formula} expression
}
\examples{
df <- iris[c(1, 2, 3, 4)]
ggfortify:::parse.formula(y ~ x)
}

