% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.label}
\alias{plot.label}
\title{Attach label to \code{ggplot2::ggplot}}
\usage{
\method{plot}{label}(p, data, flag = TRUE, label = "rownames",
  colour = NULL, size = 4)
}
\arguments{
\item{p}{\code{ggplot2::ggplot} instance}

\item{data}{Data contains text label}

\item{flag}{Logical value whether to display labels}

\item{label}{Column name used for label text}

\item{colour}{Text colour for labels}

\item{size}{Text size for labels}
}
\value{
ggplot
}
\description{
Attach label to \code{ggplot2::ggplot}
}

