% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotlib.R
\name{plot_confint}
\alias{plot_confint}
\title{Attach confidence interval to \code{ggplot2::ggplot}}
\usage{
plot_confint(p, data = NULL, lower = "lower", upper = "upper",
  conf.int = TRUE, colour = "#0000FF", linetype = "none",
  fill = "#000000", alpha = 0.3)
}
\arguments{
\item{p}{\code{ggplot2::ggplot} instance}

\item{data}{Data contains lower and upper confidence intervals}

\item{lower}{Column name for lower confidence interval}

\item{upper}{Column name for upper confidence interval}

\item{conf.int}{Logical flag indicating whether to plot confidence intervals}

\item{colour}{Line colour for confidence intervals}

\item{linetype}{Line type for confidence intervals}

\item{fill}{Fill colour for confidence intervals}

\item{alpha}{Alpha for confidence intervals}
}
\value{
ggplot
}
\description{
Attach confidence interval to \code{ggplot2::ggplot}
}
\examples{
d <- fortify(stats::acf(AirPassengers, plot = FALSE))
p <- ggplot(data = d, mapping = aes(x = Lag))
ggfortify:::plot_confint(p, data = d)
}

