% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{plot_label}
\alias{plot_label}
\title{Attach label to \code{ggplot2::ggplot}}
\usage{
plot_label(
  p,
  data,
  x = NULL,
  y = NULL,
  label = TRUE,
  label.label = "rownames",
  label.colour = NULL,
  label.alpha = NULL,
  label.size = NULL,
  label.angle = NULL,
  label.family = NULL,
  label.fontface = NULL,
  label.lineheight = NULL,
  label.hjust = NULL,
  label.vjust = NULL,
  label.repel = FALSE,
  label.show.legend = NA,
  label.position = "identity"
)
}
\arguments{
\item{p}{\code{ggplot2::ggplot} instance}

\item{data}{Data contains text label}

\item{x}{x coordinates for label}

\item{y}{y coordinates for label}

\item{label}{Logical value whether to display labels}

\item{label.label}{Column name used for label text}

\item{label.colour}{Colour for text labels}

\item{label.alpha}{Alpha for text labels}

\item{label.size}{Size for text labels}

\item{label.angle}{Angle for text labels}

\item{label.family}{Font family for text labels}

\item{label.fontface}{Fontface for text labels}

\item{label.lineheight}{Lineheight for text labels}

\item{label.hjust}{Horizontal adjustment for text labels}

\item{label.vjust}{Vertical adjustment for text labels}

\item{label.repel}{Logical flag indicating whether to use \code{ggrepel}, enabling this may take some time for plotting}

\item{label.show.legend}{Logical value indicating whether to show the legend of the text labels}

\item{label.position}{Character or a position function}
}
\value{
ggplot
}
\description{
Attach label to \code{ggplot2::ggplot}
}
