% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotlib.R
\docType{methods}
\name{+,ggmultiplot,ANY-method}
\alias{+,ggmultiplot,ANY-method}
\title{Generic add operator for \code{ggmultiplot}}
\usage{
\S4method{+}{ggmultiplot,ANY}(e1, e2)
}
\arguments{
\item{object}{\code{ggmultiplot}}

\item{g}{}
}
\value{
\code{ggmultiplot}
}
\description{
Generic add operator for \code{ggmultiplot}
}

