% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{post_fortify}
\alias{post_fortify}
\title{Post process for fortify.}
\usage{
post_fortify(data, klass = NULL)
}
\arguments{
\item{data}{data.frame}

\item{klass}{instance to be added as base_class attr, should be original model before fortified}
}
\value{
data.frame
}
\description{
Post process for fortify.
}
