% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotlib.R
\name{print.ggmultiplot}
\alias{print.ggmultiplot}
\title{Generic print function for \code{ggmultiplot}}
\usage{
\method{print}{ggmultiplot}(p)
}
\arguments{
\item{p}{\code{ggmultiplot}}
}
\description{
Generic print function for \code{ggmultiplot}
}

