% Generated by roxygen2 (4.0.2): do not edit by hand
\name{residuals.ar}
\alias{residuals.ar}
\title{Calcurate residuals for \code{stats::ar}}
\usage{
\method{residuals}{ar}(object, ...)
}
\arguments{
\item{object}{\code{stats::ar} instance.}
}
\value{
ts Residuals extracted from the object object.
}
\description{
Calcurate residuals for \code{stats::ar}
}
\examples{
fitted(ar(WWWusage))
}

