% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{support_autoplot}
\alias{support_autoplot}
\title{Check if passed object is supported by \code{ggplot2::autoplot}}
\usage{
support_autoplot(obj)
}
\arguments{
\item{obj}{object}
}
\value{
logical
}
\description{
Check if passed object is supported by \code{ggplot2::autoplot}
}
