% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unscale}
\alias{unscale}
\title{Backtransform \code{scale}-ed object}
\usage{
unscale(data, center = NULL, scale = NULL)
}
\arguments{
\item{data}{Scaled data}

\item{center}{Centered vector}

\item{scale}{Scale vector}
}
\value{
data.frame
}
\description{
Backtransform \code{scale}-ed object
}
\examples{
df <- iris[c(1, 2, 3, 4)]
ggfortify::unscale(base::scale(df))
}

