% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.acf}
\alias{autoplot.acf}
\title{Autoplot \code{stats::acf}.}
\usage{
\method{autoplot}{acf}(data, conf.int = TRUE, conf.int.fill = "#000000",
  conf.int.alpha = 0.3)
}
\arguments{
\item{data}{\code{stats::acf} instance}

\item{conf.int}{Logical flag indicating whether to plot confidence intervals}

\item{conf.int.fill}{Fill colour for confidence intervals}

\item{conf.int.alpha}{Alpha for confidence intervals}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::acf}.
}
\examples{
ggplot2::autoplot(stats::acf(AirPassengers))
ggplot2::autoplot(stats::pacf(AirPassengers))
ggplot2::autoplot(stats::ccf(AirPassengers, AirPassengers))
}

