% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.acf}
\alias{autoplot.acf}
\title{Autoplot \code{stats::acf}.}
\usage{
\method{autoplot}{acf}(data)
}
\arguments{
\item{data}{\code{stats::acf} instance}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::acf}.
}
\examples{
ggplot2::autoplot(stats::acf(AirPassengers))
ggplot2::autoplot(stats::pacf(AirPassengers))
ggplot2::autoplot(stats::ccf(AirPassengers, AirPassengers))
}

