% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_changepoint.R
\name{autoplot.cpt}
\alias{autoplot.cpt}
\title{Autoplot \code{changepoint::cpt}.}
\usage{
\method{autoplot}{cpt}(object, is.date = NULL, cpt.colour = "#FF0000",
  cpt.linetype = "dashed", ...)
}
\arguments{
\item{object}{\code{changepoint::cpt} instance}

\item{is.date}{Logical frag indicates whether the \code{stats::ts} is date or not.
If not provided, regard the input as date when the frequency is 4 or 12.}

\item{cpt.colour}{Line colour for changepoints}

\item{cpt.linetype}{Line type for changepoints}

\item{...}{Keywords passed to autoplot.ts}
}
\value{
ggplot
}
\description{
Autoplot \code{changepoint::cpt}.
}
\examples{
library(changepoint)
ggplot2::autoplot(cpt.mean(AirPassengers))
ggplot2::autoplot(cpt.meanvar(AirPassengers))
}

