% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.cpt}
\alias{autoplot.cpt}
\title{Autoplot \code{changepoint::cpt}.}
\usage{
\method{autoplot}{cpt}(data, cpt.colour = "#FF0000",
  cpt.linetype = "dashed", ...)
}
\arguments{
\item{data}{\code{changepoint::cpt} instance}

\item{cpt.colour}{Line colour for \code{changepoint::cpt}}

\item{cpt.linetype}{Line type for \code{changepoint::cpt}}

\item{...}{Keywords passed to autoplot.ts}
}
\value{
ggplot
}
\description{
Autoplot \code{changepoint::cpt}.
}
\examples{
ggplot2::autoplot(changepoint::cpt.mean(AirPassengers))
ggplot2::autoplot(changepoint::cpt.meanvar(AirPassengers))
}

