% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_stats_density.R
\name{autoplot.density}
\alias{autoplot.density}
\title{Autoplot \code{stats::density}}
\usage{
\method{autoplot}{density}(object, p = NULL, colour = "#000000",
  linetype = NULL, fill = NULL, alpha = NULL, ...)
}
\arguments{
\item{object}{\code{stats::density} instance}

\item{p}{\code{ggplot2::ggplot} instance to plot}

\item{colour}{Line colour}

\item{linetype}{Line type}

\item{fill}{Fill colour}

\item{alpha}{Alpha}

\item{...}{other arguments passed to PDC/CDF func}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::density}
}
\examples{
autoplot(stats::density(stats::rnorm(1:50)))
autoplot(stats::density(stats::rnorm(1:50)), fill = 'blue')
}

