% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.forecast}
\alias{autoplot.forecast}
\title{Autoplot \code{forecast::forecast}.}
\usage{
\method{autoplot}{forecast}(forecast.data)
}
\arguments{
\item{forecast.data}{\code{forecast::forecast} instance}
}
\value{
data.frame
}
\description{
Autoplot \code{forecast::forecast}.
}
\examples{
d.arima <- forecast::auto.arima(AirPassengers)
ggplot2::autoplot(forecast::forecast(d.arima, h = 50))
ggplot2::autoplot(forecast::forecast(d.arima, level = c(95), h = 50))
ggplot2::autoplot(forecast::forecast(d.arima, level = c(80), h = 50))
}

