% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.forecast}
\alias{autoplot.forecast}
\title{Autoplot \code{forecast::forecast}.}
\usage{
\method{autoplot}{forecast}(data, conf.int = TRUE)
}
\arguments{
\item{data}{\code{forecast::forecast} instance}

\item{conf.int}{Logical flag indicating whether to plot confidence intervals}
}
\value{
ggplot
}
\description{
Autoplot \code{forecast::forecast}.
}
\examples{
d.arima <- forecast::auto.arima(AirPassengers)
ggplot2::autoplot(forecast::forecast(d.arima, h = 50))
ggplot2::autoplot(forecast::forecast(d.arima, level = c(95), h = 50))
ggplot2::autoplot(forecast::forecast(d.arima, level = c(80), h = 50))
}

