% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_cluster.R
\name{autoplot.kmeans}
\alias{autoplot.kmeans}
\title{Autoplot \code{stats::kmeans}, \code{cluster::clara}, \code{cluster::fanny} and
\code{cluster::pam}.}
\usage{
\method{autoplot}{kmeans}(data, original = NULL, ...)
}
\arguments{
\item{data}{Clustered instance}

\item{original}{Original data used for clustering. Mandatory for \code{stats::kmeans}.}

\item{...}{Options supported in \code{autoplot::prcomp}}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::kmeans}, \code{cluster::clara}, \code{cluster::fanny} and
\code{cluster::pam}.
}
\examples{
df <- iris[-5]
ggplot2::autoplot(stats::kmeans(df, 3), original = iris)
ggplot2::autoplot(cluster::clara(df, 3), label = TRUE)
ggplot2::autoplot(cluster::fanny(df, 3))
ggplot2::autoplot(cluster::fanny(df, 3), frame = TRUE)
ggplot2::autoplot(cluster::pam(df, 3), original = iris)
ggplot2::autoplot(cluster::pam(df, 3), original = iris, frame = TRUE, frame.type = 't')
}

