% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_cluster.R
\name{autoplot.kmeans}
\alias{autoplot.kmeans}
\title{Autoplot \code{stats::kmeans}, \code{cluster::clara}, \code{cluster::fanny} and
\code{cluster::pam}.}
\usage{
\method{autoplot}{kmeans}(object, data = NULL, original = NULL,
  colour = "cluster", ...)
}
\arguments{
\item{object}{Clustered instance}

\item{data}{Original data used for clustering. Mandatory for \code{stats::kmeans}.}

\item{original}{(Deprecated) use data}

\item{colour}{line colour for points}

\item{...}{other arguments passed to \code{autoplot::prcomp}}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::kmeans}, \code{cluster::clara}, \code{cluster::fanny} and
\code{cluster::pam}.
}
\examples{
autoplot(stats::kmeans(iris[-5], 3), data = iris)
autoplot(cluster::clara(iris[-5], 3), label = TRUE)
autoplot(cluster::fanny(iris[-5], 3))
autoplot(cluster::fanny(iris[-5], 3), frame = TRUE)
autoplot(cluster::pam(iris[-5], 3), data = iris, colour = 'Species')
autoplot(cluster::pam(iris[-5], 3), data = iris, frame = TRUE, frame.type = 't')
}

