% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.kmeans}
\alias{autoplot.kmeans}
\title{Autoplot \code{stats::kmeans}.}
\usage{
\method{autoplot}{kmeans}(data, original = NULL, ...)
}
\arguments{
\item{data}{\code{stats::kmeans} instance}

\item{original}{Original data used for K-means. Mandatory for plotting.}

\item{...}{Options supported in \code{autoplot::prcomp}}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::kmeans}.
}
\examples{
df <- iris[-5]
ggplot2::autoplot(stats::kmeans(df, 3), original = iris)
}

