% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_fortify_list.R
\name{autoplot.list}
\alias{autoplot.list}
\title{Autoplot list}
\usage{
\method{autoplot}{list}(object, data = NULL, nrow = NULL, ncol = NULL, scales = "free_y", ...)
}
\arguments{
\item{object}{\code{list} instance}

\item{data}{original dataset, if needed}

\item{nrow}{Number of facet/subplot rows}

\item{ncol}{Number of facet/subplot columns}

\item{scales}{Scale value passed to \code{ggplot2}}

\item{...}{other arguments passed to methods}
}
\value{
ggplot
}
\description{
Autoplot list
}
