% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_base.R
\name{autoplot.matrix}
\alias{autoplot.matrix}
\title{Plot \code{base::matrix}}
\usage{
\method{autoplot}{matrix}(object, original = NULL, fill = "#0000FF",
  scale = NULL, colour = NULL, label = FALSE, label.colour = colour,
  label.size = 4, geom = "tile", ...)
}
\arguments{
\item{object}{\code{base::matrix} instance}

\item{original}{Combined to data by column if provided. Intended to be used for stat functions which
returns not containing original data.}

\item{fill}{Fill (high) colour. Ignored if scale keyword is passed. ('tile' Only)}

\item{scale}{\code{ggplot2::scale} instance to plot. ('tile' Only)}

\item{colour}{Column name string to specify colorize points ('point' Only)}

\item{label}{Logical value whether to display data labels ('point' Only)}

\item{label.colour}{Text colour for data labels ('point' Only)}

\item{label.size}{Text size for data labels ('point' Only)}

\item{geom}{Geometric string for plotting. 'tile' or 'point'.}

\item{...}{other arguments passed to methods}
}
\value{
ggplot
}
\description{
Plot \code{base::matrix}
}
\examples{
autoplot(matrix(rnorm(20), nc = 5))
autoplot(matrix(rnorm(20), nc = 5), fill = 'red')
autoplot(matrix(rnorm(20), nc = 5),
         scale = scale_fill_gradient(low = 'red', high = 'blue'))
autoplot(matrix(rnorm(20), nc = 2), geom = 'point')
}

