% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.pca_common}
\alias{autoplot.pca_common}
\title{Autoplot PCA-likes.}
\usage{
\method{autoplot}{pca_common}(data, original = NULL, colour = NULL,
  label = FALSE, label.size = 4, loadings = FALSE,
  loadings.colour = "#FF0000", loadings.label = FALSE,
  loadings.label.colour = "#FF0000", loadings.label.size = 4)
}
\arguments{
\item{data}{PCA-like instance}

\item{original}{Joined to fitting result if provided.}

\item{colour}{Column name string to specify colorize points}

\item{label}{Logical value whether to display data labels}

\item{label.size}{Text size for data labels}

\item{loadings}{Logical value whether to display loadings arrows}

\item{loadings.colour}{Point colour for data}

\item{loadings.label}{Logical value whether to display loadings labels}

\item{loadings.label.colour}{Text colour for loadings labels}

\item{loadings.label.size}{Text size for loadings labels}
}
\value{
ggplot
}
\description{
Autoplot PCA-likes.
}
\examples{
df <- iris[c(1, 2, 3, 4)]
ggplot2::autoplot(stats::prcomp(df))
ggplot2::autoplot(stats::prcomp(df), original = iris)
ggplot2::autoplot(stats::prcomp(df), original = iris, colour = 'Species')
ggplot2::autoplot(stats::prcomp(df), label = TRUE, loadings = TRUE, loadings.label = TRUE)

ggplot2::autoplot(stats::princomp(df))
ggplot2::autoplot(stats::princomp(df), original = iris)
ggplot2::autoplot(stats::princomp(df), original = iris, colour = 'Species')
ggplot2::autoplot(stats::princomp(df), label = TRUE, loadings = TRUE, loadings.label = TRUE)

d.factanal <- stats::factanal(state.x77, factors = 3, scores = 'regression')
ggplot2::autoplot(d.factanal)
ggplot2::autoplot(d.factanal, original = state.x77, colour = 'Income')
ggplot2::autoplot(d.factanal, label = TRUE, loadings = TRUE, loadings.label = TRUE)
}

