% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_performance.R
\name{autoplot.performance}
\alias{autoplot.performance}
\title{Autoplot \code{ROCR::performance}}
\usage{
\method{autoplot}{performance}(object, p = NULL, bins = 5, ...)
}
\arguments{
\item{object}{\code{ROCR::performance} instance}

\item{p}{\code{ggplot2::ggplot} instances}

\item{bins}{If \code{object} represents a measure whose value is just a
scalar (e.g. \code{performance(predObj, 'auc')}), a histogram will be
plotted of this scalar's values for different runs. \code{bins}
is the number of bins for this histogram.}

\item{...}{other arguments passed to methods}
}
\value{
ggplot
}
\description{
Autoplot \code{ROCR::performance}
}
