% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.prcomp}
\alias{autoplot.prcomp}
\title{Autoplot \code{stats::prcomp}.}
\usage{
\method{autoplot}{prcomp}(data, original = NULL)
}
\arguments{
\item{data}{\code{stats::prcomp} instance}

\item{original}{Joined to PCA result if provided. Intended to be used for attaching
non-numeric values original data has. Numeric values are automatically attached.}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::prcomp}.
}
\examples{
df <- iris[c(1, 2, 3, 4)]
ggplot2::autoplot(stats::prcomp(df))
ggplot2::autoplot(stats::prcomp(df), original = iris)
}

