% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.prcomp}
\alias{autoplot.prcomp}
\title{Autoplot \code{stats::prcomp}.}
\usage{
\method{autoplot}{prcomp}(data)
}
\arguments{
\item{data}{\code{stats::prcomp} instance}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::prcomp}.
}
\examples{
df <- iris[c(1, 2, 3, 4)]
ggplot2::autoplot(stats::prcomp(df))
}

