% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.princomp}
\alias{autoplot.princomp}
\title{Autoplot \code{stats::princomp}.}
\usage{
\method{autoplot}{princomp}(data, original = NULL)
}
\arguments{
\item{data}{\code{stats::princomp} instance}

\item{original}{Joined to PCA result if provided. Intended to be used for attaching
non-numeric values original data has. Numeric values are automatically attached.}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::princomp}.
}
\examples{
df <- iris[c(1, 2, 3, 4)]
ggplot2::autoplot(stats::princomp(df))
ggplot2::autoplot(stats::princomp(df), original = iris)
}

