% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.princomp}
\alias{autoplot.princomp}
\title{Autoplot \code{stats::princomp}.}
\usage{
\method{autoplot}{princomp}(data)
}
\arguments{
\item{data}{\code{stats::princomp} instance}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::princomp}.
}
\examples{
df <- iris[c(1, 2, 3, 4)]
ggplot2::autoplot(stats::princomp(df))
}

