% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_stats.R
\name{autoplot.spec}
\alias{autoplot.spec}
\title{Autoplot \code{stats::spec}.}
\usage{
\method{autoplot}{spec}(object, ...)
}
\arguments{
\item{object}{\code{stats::spec} instance}

\item{...}{other arguments passed to specific methods}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::spec}.
}
\examples{
ggplot2::autoplot(stats::spec.ar(AirPassengers))
ggplot2::autoplot(stats::spec.pgram(AirPassengers))
}

