% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.stl}
\alias{autoplot.stl}
\alias{fortify.decomposed.ts}
\title{Autoplot \code{stats::stl} and \code{stats::decomposed.ts}}
\usage{
\method{autoplot}{stl}(data, scales = "free_y", ts.colour = "#000000",
  ts.linetype = "solid")
}
\arguments{
\item{data}{\code{stats::stl} or \code{stats::decomposed.ts} instance}

\item{scales}{Scale value passed to \code{ggplot2}}

\item{ts.colour}{Line colour for time-series}

\item{ts.linetype}{Line type for time-series}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::stl} and \code{stats::decomposed.ts}
}
\examples{
ggplot2::autoplot(stats::stl(UKgas, s.window = 'periodic'))
ggplot2::autoplot(stats::decompose(UKgas))
}

