% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.survfit}
\alias{autoplot.survfit}
\title{Autoplot \code{survival::survfit}.}
\usage{
\method{autoplot}{survfit}(survfit.data)
}
\arguments{
\item{survfit.data}{\code{survival::survfit} instance}
}
\value{
data.frame
}
\description{
Autoplot \code{survival::survfit}.
}
\examples{
d.survfit <- survival::survfit(survival::Surv(time, status) ~ sex, data = survival::lung)
ggplot2::autoplot(d.survfit)
}

