% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_surv.R
\name{autoplot.survfit}
\alias{autoplot.survfit}
\alias{autoplot.survfit.cox}
\title{Autoplot \code{survival::survfit}.}
\usage{
\method{autoplot}{survfit}(object, surv.colour = "#000000",
  surv.linetype = "solid", conf.int = TRUE, conf.int.colour = "#0000FF",
  conf.int.linetype = "none", conf.int.fill = "#000000",
  conf.int.alpha = 0.3, censor = TRUE, censor.shape = "+",
  censor.size = 3, ...)
}
\arguments{
\item{object}{\code{survival::survfit} instance}

\item{surv.colour}{Line colour for survival curve (used when data doesn't have \code{strata})}

\item{surv.linetype}{Line type for survival curve}

\item{conf.int}{Logical flag indicating whether to plot confidence intervals}

\item{conf.int.colour}{Line colour for confidence intervals}

\item{conf.int.linetype}{Line type for confidence intervals}

\item{conf.int.fill}{Fill colour for confidence intervals}

\item{conf.int.alpha}{Alpha for confidence intervals}

\item{censor}{Logical flag indicating whether to plot censors}

\item{censor.shape}{Shape for censors}

\item{censor.size}{Size for censors}

\item{...}{other arguments passed to methods}
}
\value{
ggplot
}
\description{
Autoplot \code{survival::survfit}.
}
\examples{
d.survfit <- survival::survfit(survival::Surv(time, status) ~ sex, data = survival::lung)
ggplot2::autoplot(d.survfit)
ggplot2::autoplot(d.survfit, conf.int = FALSE, censor = FALSE)

d.coxph <- survival::coxph(survival::Surv(time, status) ~ sex, data = survival::lung)
ggplot2::autoplot(survival::survfit(d.coxph))
}

