% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.survfit}
\alias{autoplot.survfit}
\alias{autoplot.survfit.cox}
\title{Autoplot \code{survival::survfit}.}
\usage{
\method{autoplot}{survfit}(data, surv.colour = "#000000",
  surv.linetype = "solid", conf.int = TRUE, conf.int.fill = "#000000",
  conf.int.alpha = 0.3, censor = TRUE, censor.shape = "+",
  censor.size = 3)
}
\arguments{
\item{data}{\code{survival::survfit} instance}

\item{surv.colour}{Line colour for survival curve (used when data doesn't have \code{strata})}

\item{surv.linetype}{Line type for survival curve}

\item{conf.int}{Logical flag indicating whether to plot confidence intervals}

\item{conf.int.fill}{Fill colour for confidence intervals}

\item{conf.int.alpha}{Alpha for confidence intervals}

\item{censor}{Logical flag indicating whether to plot censors}

\item{censor.shape}{Shape for censors}

\item{censor.size}{Size for censors}
}
\value{
ggplot
}
\description{
Autoplot \code{survival::survfit}.
}
\examples{
d.survfit <- survival::survfit(survival::Surv(time, status) ~ sex, data = survival::lung)
ggplot2::autoplot(d.survfit)
ggplot2::autoplot(d.survfit, conf.int = FALSE, censor = FALSE)
}

