% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.survfit}
\alias{autoplot.survfit}
\title{Autoplot \code{survival::survfit}.}
\usage{
\method{autoplot}{survfit}(data, conf.int = TRUE, censor = TRUE)
}
\arguments{
\item{data}{\code{survival::survfit} instance}

\item{conf.int}{Logical flag indicating whether to plot confidence intervals}

\item{conf.int}{Logical flag indicating whether to plot censors}
}
\value{
ggplot
}
\description{
Autoplot \code{survival::survfit}.
}
\examples{
d.survfit <- survival::survfit(survival::Surv(time, status) ~ sex, data = survival::lung)
ggplot2::autoplot(d.survfit)
ggplot2::autoplot(d.survfit, conf.int = FALSE, censor = FALSE)
}

