% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.ts}
\alias{autoplot.ts}
\alias{autoplot.xts}
\title{Autoplot time-series-like.}
\usage{
\method{autoplot}{ts}(data, scales = "free_y", facet = TRUE,
  ts.colour = "#000000", ts.linetype = "solid")
}
\arguments{
\item{data}{time-series-like instance}

\item{scales}{Scale value passed to \code{ggplot2}}

\item{facet}{Logical value to specify use facets for multivariate time-series.}

\item{ts.colour}{Line colour for time-series}

\item{ts.linetype}{Line type for time-series}
}
\value{
ggplot
}
\description{
Autoplot time-series-like.
}
\examples{
data(Canada, package = 'vars')
ggplot2::autoplot(AirPassengers)
ggplot2::autoplot(UKgas)
ggplot2::autoplot(Canada)
ggplot2::autoplot(Canada, facet = FALSE)

library(zoo)
ggplot2::autoplot(xts::as.xts(AirPassengers))
ggplot2::autoplot(xts::as.xts(UKgas))
ggplot2::autoplot(xts::as.xts(Canada))

ggplot2::autoplot(timeSeries::as.timeSeries(AirPassengers))
ggplot2::autoplot(timeSeries::as.timeSeries(Canada))

its <- tseries::irts(cumsum(rexp(10, rate = 0.1)), matrix(rnorm(20), nrow=2))
ggplot2::autoplot(its)

ggplot2::autoplot(stats::arima(UKgas))
ggplot2::autoplot(stats::ar(AirPassengers))
}

