% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.ts}
\alias{autoplot.ts}
\title{Autoplot \code{stats::ts}.}
\usage{
\method{autoplot}{ts}(data, scales = "free_y")
}
\arguments{
\item{data}{\code{stats::ts} instance}

\item{scales}{Scale value passed to \code{ggplot2}}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::ts}.
}
\examples{
data(Canada, package = 'vars')
ggplot2::autoplot(AirPassengers)
ggplot2::autoplot(Canada)
}

