% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.ts}
\alias{autoplot.decomposed.ts}
\alias{autoplot.irts}
\alias{autoplot.stl}
\alias{autoplot.timeSeries}
\alias{autoplot.ts}
\alias{autoplot.xts}
\title{Autoplot time-series-like.}
\usage{
\method{autoplot}{ts}(data, columns = NULL, group = NULL, is.date = NULL,
  index.name = "Index", p = NULL, scales = "free_y", facet = TRUE,
  ts.geom = "line", ts.colour = "#000000", ts.linetype = "solid",
  xlab = "", ylab = "")
}
\arguments{
\item{data}{time-series-like instance}

\item{columns}{Character vector specifies target column name(s)}

\item{group}{Character vector specifies grouping}

\item{is.date}{Logical frag indicates whether the \code{stats::ts} is date or not.
If not provided, regard the input as date when the frequency is 4 or 12.}

\item{index.name}{Specify column name for time series index when passing \code{data.frame} via data.}

\item{p}{\code{ggplot2::ggplot} instance}

\item{scales}{Scale value passed to \code{ggplot2}}

\item{facet}{Logical value to specify use facets for multivariate time-series}

\item{ts.geom}{Geometric string for colour for time-series. 'line' or 'bar'.}

\item{ts.colour}{Line colour for time-series}

\item{ts.linetype}{Line type for time-series}

\item{xlab}{Character vector or expression for x axis label}

\item{ylab}{Character vector or expression for y axis label}
}
\value{
ggplot
}
\description{
Autoplot time-series-like.
}
\examples{
data(Canada, package = 'vars')
ggplot2::autoplot(AirPassengers)
ggplot2::autoplot(UKgas, ts.geom = 'bar')
ggplot2::autoplot(Canada)
ggplot2::autoplot(Canada, columns = 'e', is.date = TRUE)
ggplot2::autoplot(Canada, facet = FALSE)

library(zoo)
ggplot2::autoplot(xts::as.xts(AirPassengers))
ggplot2::autoplot(xts::as.xts(UKgas))
ggplot2::autoplot(xts::as.xts(Canada))

ggplot2::autoplot(timeSeries::as.timeSeries(AirPassengers))
ggplot2::autoplot(timeSeries::as.timeSeries(Canada))

its <- tseries::irts(cumsum(rexp(10, rate = 0.1)), matrix(rnorm(20), ncol=2))
ggplot2::autoplot(its)

ggplot2::autoplot(stats::stl(UKgas, s.window = 'periodic'))
ggplot2::autoplot(stats::decompose(UKgas))
}

