% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.ts}
\alias{autoplot.ts}
\title{Autoplot \code{stats::ts}.}
\usage{
\method{autoplot}{ts}(data, scales = "free_y", facet = TRUE,
  ts.colour = "#000000", ts.linetype = "solid")
}
\arguments{
\item{data}{\code{stats::ts} instance}

\item{scales}{Scale value passed to \code{ggplot2}}

\item{facet}{Logical value to specify use facets for multivariate \code{stats::ts}.}

\item{ts.colour}{Line colour for \code{stats::ts}}

\item{ts.linetype}{Line type for \code{stats::ts}}
}
\value{
ggplot
}
\description{
Autoplot \code{stats::ts}.
}
\examples{
data(Canada, package = 'vars')
ggplot2::autoplot(AirPassengers)
ggplot2::autoplot(Canada)
ggplot2::autoplot(Canada, facet = FALSE)
}

