% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.tsmodel}
\alias{autoplot.HoltWinters}
\alias{autoplot.ar}
\alias{autoplot.fGARCH}
\alias{autoplot.fracdiff}
\alias{autoplot.nnetar}
\alias{autoplot.tsmodel}
\title{Autoplot time series models (like AR, ARIMA).}
\usage{
\method{autoplot}{tsmodel}(data, original = NULL, is.date = NULL,
  fitted.colour = "#FF0000", fitted.linetype = "solid", ...)
}
\arguments{
\item{data}{\code{stats::ar}, \code{stats::Arima}, \code{stats::HoltWinters},
\code{fracdiff::fracdiff}, \code{forecast::nnetar} or \code{fGarch::fGARCH}instance}

\item{original}{Original data for \code{stats::ar}, \code{stats::Arima}.
Not used in other models.
If not provided, try to retrieve from current environment using variable name.}

\item{is.date}{Logical frag indicates whether the \code{stats::ts} is date or not.
If not provided, regard the input as date when the frequency is 4 or 12.}

\item{fitted.colour}{Line colour for fitted time-series}

\item{fitted.linetype}{Line type for fitted time-series}

\item{...}{Keywords passed to autoplot.ts}
}
\value{
ggplot
}
\description{
Autoplot time series models (like AR, ARIMA).
}
\examples{
ggplot2::autoplot(stats::ar(AirPassengers))
ggplot2::autoplot(stats::arima(UKgas), original = UKgas)
ggplot2::autoplot(forecast::arfima(AirPassengers))
ggplot2::autoplot(forecast::nnetar(UKgas), is.date = FALSE)
ggplot2::autoplot(stats::HoltWinters(USAccDeaths))
}

