% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.varprd}
\alias{autoplot.varprd}
\title{Autoplot \code{vars::varprd}.}
\usage{
\method{autoplot}{varprd}(data, is.date = NULL, ts.connect = TRUE,
  scales = "free_y", predict.colour = "#0000FF",
  predict.linetype = "solid", conf.int = TRUE,
  conf.int.colour = "#0000FF", conf.int.linetype = "none",
  conf.int.fill = "#000000", conf.int.alpha = 0.3, ...)
}
\arguments{
\item{data}{\code{vars::varpred} instance}

\item{is.date}{Logical frag indicates whether the \code{stats::ts} is date or not.
If not provided, regard the input as date when the frequency is 4 or 12.}

\item{ts.connect}{Logical frag indicates whether connects original time-series and predicted values}

\item{scales}{Scale value passed to \code{ggplot2}}

\item{predict.colour}{Line colour for predicted time-series}

\item{predict.linetype}{Line type for predicted time-series}

\item{conf.int}{Logical flag indicating whether to plot confidence intervals}

\item{conf.int.colour}{Line colour for confidence intervals}

\item{conf.int.linetype}{Line type for confidence intervals}

\item{conf.int.fill}{Fill colour for confidence intervals}

\item{conf.int.alpha}{Alpha for confidence intervals}

\item{...}{Keywords passed to autoplot.ts}
}
\value{
ggplot
}
\description{
Autoplot \code{vars::varprd}.
}
\examples{
data(Canada, package = 'vars')
d.vselect <- vars::VARselect(Canada, lag.max = 5, type = 'const')$selection[1]
d.var <- vars::VAR(Canada, p = d.vselect, type = 'const')
ggplot2::autoplot(stats::predict(d.var, n.ahead = 50), is.date = TRUE)
ggplot2::autoplot(stats::predict(d.var, n.ahead = 50), conf.int = FALSE)
}

