% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cbind_wraps}
\alias{cbind_wraps}
\title{Wrapper for cbind}
\usage{
cbind_wraps(df1, df2)
}
\arguments{
\item{df1}{1st data}

\item{df2}{2nd data}
}
\value{
list
}
\description{
Wrapper for cbind
}
\examples{
ggfortify:::cbind_wraps(iris[1:2], iris[3:5])
}
