% Generated by roxygen2 (4.0.2): do not edit by hand
\name{confint.acf}
\alias{confint.acf}
\title{Calcurate confidence interval for \code{stats::acf}}
\usage{
\method{confint}{acf}(x, ci = 0.95, ci.type = "white")
}
\arguments{
\item{x}{\code{stats::acf} instance}

\item{ci}{Float value for confidence interval}

\item{ci.type}{"white" or "ma"}
}
\value{
vector
}
\description{
Calcurate confidence interval for \code{stats::acf}
}
\examples{
confint(acf(AirPassengers))
confint(acf(AirPassengers), ci.type = 'ma')
}

