% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{deprecate.warning}
\alias{deprecate.warning}
\title{Show deprecate warning}
\usage{
deprecate.warning(old.kw, new.kw)
}
\arguments{
\item{old.kw}{Keyword being deprecated}

\item{new.kw}{Keyword being replaced}
}
\description{
Show deprecate warning
}
\examples{
ggfortify:::deprecate.warning('old', 'new')
}
