% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslib.R
\name{fitted.ar}
\alias{fitted.ar}
\title{Calculate fitted values for \code{stats::ar}}
\usage{
\method{fitted}{ar}(object, ...)
}
\arguments{
\item{object}{\code{stats::ar} instance}

\item{...}{other keywords}
}
\value{
ts An time series of the one-step forecasts
}
\description{
Calculate fitted values for \code{stats::ar}
}
\examples{
fitted(ar(WWWusage))
}
