% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.acf}
\alias{fortify.acf}
\title{Convert \code{stats::acf} to data.frame.}
\usage{
\method{fortify}{acf}(acf.data)
}
\arguments{
\item{acf.data}{\code{stats::acf} instance}
}
\value{
data.frame
}
\description{
Convert \code{stats::acf} to data.frame.
}
\examples{
ggplot2::fortify(stats::acf(AirPassengers))
ggplot2::fortify(stats::pacf(AirPassengers))
ggplot2::fortify(stats::ccf(AirPassengers, AirPassengers))
}

