% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.Arima}
\alias{fortify.Arima}
\alias{fortify.HoltWinters}
\alias{fortify.ar}
\alias{fortify.fracdiff}
\alias{fortify.nnetar}
\title{Convert AR, ARIMA, ARFIMA-like to data.frame.}
\usage{
\method{fortify}{Arima}(data)
}
\arguments{
\item{data}{\code{stats::ar}, \code{stats::arima}, \code{fracdiff::fracdiff}
or \code{forecast::nnetar}instance}
}
\value{
data.frame
}
\description{
Convert AR, ARIMA, ARFIMA-like to data.frame.
}
\examples{
ggplot2::fortify(stats::ar(AirPassengers))
ggplot2::fortify(stats::arima(UKgas))
ggplot2::fortify(forecast::auto.arima(austres))
ggplot2::fortify(forecast::arfima(AirPassengers))
ggplot2::fortify(forecast::nnetar(UKgas))
ggplot2::fortify(stats::HoltWinters(USAccDeaths))
}

