% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.cpt}
\alias{fortify.cpt}
\title{Convert \code{changepoint::cpt} to data.frame.}
\usage{
\method{fortify}{cpt}(data, keep.original = TRUE)
}
\arguments{
\item{data}{\code{chantepoint::cpt} instance}

\item{keep.original}{Logical value whether to keep original \code{stats::ts}}
}
\value{
data.frame
}
\description{
Convert \code{changepoint::cpt} to data.frame.
}
\examples{
ggplot2::fortify(changepoint::cpt.mean(AirPassengers))
ggplot2::fortify(changepoint::cpt.var(AirPassengers))
ggplot2::fortify(changepoint::cpt.meanvar(AirPassengers))
}

