% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_stats_density.R
\name{fortify.density}
\alias{fortify.density}
\title{Convert \code{stats::density} to \code{data.frame}.}
\usage{
\method{fortify}{density}(model, data, ...)
}
\arguments{
\item{model}{\code{stats::density} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{stats::density} to \code{data.frame}.
}
\examples{
ggplot2::fortify(stats::density(stats::rnorm(1:50)))
}

