% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_stats_density.R
\name{fortify.density}
\alias{fortify.density}
\title{Convert \code{stats::density} to \code{data.frame}}
\usage{
\method{fortify}{density}(model, data = NULL, ...)
}
\arguments{
\item{model}{\code{stats::density} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{stats::density} to \code{data.frame}
}
\examples{
fortify(stats::density(stats::rnorm(1:50)))
}
