% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_stats.R
\name{fortify.dist}
\alias{fortify.dist}
\title{Convert \code{stats::dist} to data.frame.}
\usage{
\method{fortify}{dist}(data)
}
\arguments{
\item{data}{\code{stats::dist} instance}
}
\value{
data.frame
}
\description{
Convert \code{stats::dist} to data.frame.
}
\examples{
ggplot2::fortify(eurodist)
}

