% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_stats.R
\name{fortify.dist}
\alias{fortify.dist}
\title{Convert \code{stats::dist} to \code{data.frame}}
\usage{
\method{fortify}{dist}(model, data = NULL, ...)
}
\arguments{
\item{model}{\code{stats::dist} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{stats::dist} to \code{data.frame}
}
\examples{
fortify(eurodist)
}
