% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_stats.R
\name{fortify.factanal}
\alias{fortify.factanal}
\title{Convert \code{stats::factanal} to data.frame.}
\usage{
\method{fortify}{factanal}(model, data = NULL, original = NULL, ...)
}
\arguments{
\item{model}{\code{stats::factanal} instance}

\item{data}{original dataset, if needed}

\item{original}{(Deprecated) use data}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{stats::factanal} to data.frame.
}
\examples{
d.factanal <- stats::factanal(state.x77, factors = 3, scores = 'regression')
ggplot2::fortify(d.factanal)
ggplot2::fortify(d.factanal, original = state.x77)
}

