% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_stats.R
\name{fortify.factanal}
\alias{fortify.factanal}
\title{Convert \code{stats::factanal} to \code{data.frame}}
\usage{
\method{fortify}{factanal}(model, data = NULL, ...)
}
\arguments{
\item{model}{\code{stats::factanal} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{stats::factanal} to \code{data.frame}
}
\examples{
d.factanal <- stats::factanal(state.x77, factors = 3, scores = 'regression')
fortify(d.factanal)
fortify(d.factanal, data = state.x77)
}
