% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_cluster.R
\name{fortify.kmeans}
\alias{fortify.kmeans}
\title{Convert \code{stats::kmeans}, \code{cluster::clara}, \code{cluster::fanny} and
\code{cluster::pam} to data.frame.}
\usage{
\method{fortify}{kmeans}(data, original = NULL)
}
\arguments{
\item{data}{Clustered instance}

\item{original}{Joined to data if provided. Intended to be used for attaching
cluster labels to the original}
}
\value{
data.frame
}
\description{
Convert \code{stats::kmeans}, \code{cluster::clara}, \code{cluster::fanny} and
\code{cluster::pam} to data.frame.
}
\examples{
df <- iris[-5]
ggplot2::fortify(stats::kmeans(df, 3))
ggplot2::fortify(stats::kmeans(df, 3), original = iris)
ggplot2::fortify(cluster::clara(df, 3))
ggplot2::fortify(cluster::fanny(df, 3))
ggplot2::fortify(cluster::pam(df, 3), original = iris)
}

