% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.kmeans}
\alias{fortify.kmeans}
\title{Convert \code{stats::kmeans} to data.frame.}
\usage{
\method{fortify}{kmeans}(data, original = NULL)
}
\arguments{
\item{data}{\code{stats::kmeans} instance}

\item{original}{Joined to K-means result if provided. Intended to be used for attaching
cluster labels to the original}
}
\value{
data.frame
}
\description{
Convert \code{stats::kmeans} to data.frame.
}
\examples{
df <- iris[-5]
ggplot2::fortify(stats::kmeans(df, 3))
ggplot2::fortify(stats::kmeans(df, 3), original = iris)
}

