% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_stats.R
\name{fortify.lfda}
\alias{fortify.lfda}
\title{Convert \code{lfda::lfda} or \code{lfda::klfda} or \code{lfda::self} to \code{data.frame}}
\usage{
\method{fortify}{lfda}(model, data = NULL, ...)
}
\arguments{
\item{model}{\code{lfda::lfda} or \code{lfda::klfda} or \code{lfda::self} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{lfda::lfda} or \code{lfda::klfda} or \code{lfda::self} to \code{data.frame}
}
\examples{
\dontrun{
model <- lfda::lfda(iris[, -5], iris[, 5], 3, metric = "plain")
fortify(model)
}
}
