% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_fortify_list.R
\name{fortify.list}
\alias{fortify.list}
\title{Convert list to data.frame}
\usage{
\method{fortify}{list}(model, data = NULL, ...)
}
\arguments{
\item{model}{\code{list} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert list to data.frame
}
