% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_base.R
\name{fortify.matrix}
\alias{fortify.matrix}
\title{Convert \code{base::matrix} to data.frame.}
\usage{
\method{fortify}{matrix}(data, original = NULL, compat = FALSE)
}
\arguments{
\item{data}{\code{base::matrix} instance}

\item{original}{Combined to data by column if provided. Intended to be used for stat functions which
returns not containing original data.}

\item{compat}{Logical frag to specify the behaviour when converting matrix which has no column name.
If \code{FALSE}, result has character columns like c('1', '2', ...).
If \code{TRUE}, result has character columns like c('V1', 'V2', ...).}
}
\value{
data.frame
}
\description{
Different from \code{as.data.frame},
}
\examples{
ggplot2::fortify(matrix(1:6, nrow=2, ncol=3))
}

