% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_base.R
\name{fortify.matrix}
\alias{fortify.matrix}
\title{Convert \code{base::matrix} to \code{data.frame}}
\usage{
\method{fortify}{matrix}(model, data = NULL, compat = FALSE, ...)
}
\arguments{
\item{model}{\code{base::matrix} instance}

\item{data}{original dataset, if needed}

\item{compat}{Logical frag to specify the behaviour when converting matrix which has no column name.
If \code{FALSE}, result has character columns like c('1', '2', ...).
If \code{TRUE}, result has character columns like c('V1', 'V2', ...).}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Different from \code{as.data.frame}
}
\examples{
fortify(matrix(1:6, nrow=2, ncol=3))
}
