% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.prcomp}
\alias{fortify.prcomp}
\title{Convert \code{stats::prcomp} to data.frame.}
\usage{
\method{fortify}{prcomp}(data)
}
\arguments{
\item{data}{\code{stats::prcomp} instance}
}
\value{
data.frame
}
\description{
Convert \code{stats::prcomp} to data.frame.
}
\examples{
df <- iris[c(1, 2, 3, 4)]
ggplot2::fortify(stats::prcomp(df))
}

