% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_stats.R
\name{fortify.prcomp}
\alias{fortify.prcomp}
\alias{fortify.princomp}
\title{Convert \code{stats::prcomp}, \code{stats::princomp} to data.frame.}
\usage{
\method{fortify}{prcomp}(model, data = NULL, original = NULL, ...)
}
\arguments{
\item{model}{\code{stats::prcomp} or \code{stats::princomp} instance}

\item{data}{original dataset, if needed}

\item{original}{(Deprecated) use data}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{stats::prcomp}, \code{stats::princomp} to data.frame.
}
\examples{
df <- iris[-5]
ggplot2::fortify(stats::prcomp(df))
ggplot2::fortify(stats::prcomp(df), original = iris)

ggplot2::fortify(stats::princomp(df))
ggplot2::fortify(stats::princomp(df), original = iris)
}

