% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.spec}
\alias{fortify.spec}
\title{Convert \code{stats::spec} to data.frame.}
\usage{
\method{fortify}{spec}(data)
}
\arguments{
\item{data}{\code{stats::spec} instance}
}
\value{
data.frame
}
\description{
Convert \code{stats::spec} to data.frame.
}
\examples{
ggplot2::fortify(stats::spec.pgram(AirPassengers))
}

