% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_stats.R
\name{fortify.spec}
\alias{fortify.spec}
\title{Convert \code{stats::spec} to \code{data.frame}}
\usage{
\method{fortify}{spec}(model, data = NULL, ...)
}
\arguments{
\item{model}{\code{stats::spec} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{stats::spec} to \code{data.frame}
}
\examples{
fortify(spectrum(AirPassengers))
fortify(stats::spec.ar(AirPassengers))
fortify(stats::spec.pgram(AirPassengers))
}
