% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.stl}
\alias{fortify.stl}
\title{Convert \code{stats::stl} to data.frame.}
\usage{
\method{fortify}{stl}(stl.data)
}
\arguments{
\item{stl.data}{\code{stats::stl} instance}
}
\value{
data.frame
}
\description{
Convert \code{stats::stl} to data.frame.
}
\examples{
d.stl <- stats::stl(AirPassengers, s.window = 'periodic')
ggplot2::fortify(d.stl)
}

