% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.survfit}
\alias{fortify.survfit}
\alias{fortify.survfit.cox}
\title{Convert \code{survival::survfit} to data.frame.}
\usage{
\method{fortify}{survfit}(data)
}
\arguments{
\item{data}{\code{survival::survfit} instance}
}
\value{
data.frame
}
\description{
Convert \code{survival::survfit} to data.frame.
}
\examples{
d.survfit <- survival::survfit(survival::Surv(time, status) ~ sex, data = survival::lung)
ggplot2::fortify(d.survfit)

d.coxph <- survival::coxph(survival::Surv(time, status) ~ sex, data = survival::lung)
ggplot2::fortify(survival::survfit(d.coxph))
}

