% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_surv.R
\name{fortify.survfit}
\alias{fortify.survfit}
\alias{fortify.survfit.cox}
\title{Convert \code{survival::survfit} to data.frame.}
\usage{
\method{fortify}{survfit}(model, data = NULL, ...)
}
\arguments{
\item{model}{\code{survival::survfit} instance}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert \code{survival::survfit} to data.frame.
}
\examples{
d.survfit <- survival::survfit(survival::Surv(time, status) ~ sex, data = survival::lung)
fortify(d.survfit)

d.coxph <- survival::coxph(survival::Surv(time, status) ~ sex, data = survival::lung)
fortify(survival::survfit(d.coxph))
}

