% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.table}
\alias{fortify.table}
\title{Convert \code{base::table} to data.frame.}
\usage{
\method{fortify}{table}(data)
}
\arguments{
\item{data}{\code{base::table} instance}
}
\value{
data.frame
}
\description{
Convert \code{base::table} to data.frame.
}
\examples{
ggplot2::fortify(Titanic)
}

