% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.ts}
\alias{fortify.ts}
\title{Convert \code{stats::ts} to data.frame.}
\usage{
\method{fortify}{ts}(data)
}
\arguments{
\item{data}{\code{stats::ts} instance}
}
\value{
data.frame
}
\description{
Convert \code{stats::ts} to data.frame.
}
\examples{
ggplot2::fortify(AirPassengers)
}

