% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base_fortify_ts.R
\name{fortify.ts}
\alias{fortify.ts}
\title{Convert time-series-like to data.frame.}
\usage{
\method{fortify}{ts}(model, data = NULL, columns = NULL, is.date = NULL,
  index.name = "Index", data.name = "Data", scale = FALSE, melt = FALSE,
  ...)
}
\arguments{
\item{model}{\code{stats::ts}, \code{timeSeries::timeSeries} or \code{tseries::irts} instance}

\item{data}{original dataset, if needed}

\item{columns}{Character vector specifies target column name(s)}

\item{is.date}{Logical frag indicates whether the \code{stats::ts} is date or not.
If not provided, regard the input as date when the frequency is 4 or 12.}

\item{index.name}{Specify column name for time series index}

\item{data.name}{Specify column name for univariate time series data. Ignored in multivariate time series.}

\item{scale}{Logical flag indicating whether to perform scaling each timeseries}

\item{melt}{Logical flag indicating whether to melt each timeseries as variable}

\item{...}{other arguments passed to methods}
}
\value{
data.frame
}
\description{
Convert time-series-like to data.frame.
}
\examples{
fortify(AirPassengers)
fortify(timeSeries::as.timeSeries(AirPassengers))

its <- tseries::irts(cumsum(rexp(10, rate = 0.1)), matrix(rnorm(20), ncol=2))
fortify(its)

fortify(stats::stl(UKgas, s.window = 'periodic'))
fortify(stats::decompose(UKgas))
}

