% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.varprd}
\alias{fortify.varprd}
\title{Convert \code{vars::varprd} to data.frame.}
\usage{
\method{fortify}{varprd}(data)
}
\arguments{
\item{data}{\code{vars::varprd} instance}
}
\value{
data.frame
}
\description{
Convert \code{vars::varprd} to data.frame.
}
\examples{
data(Canada, package = 'vars')
d.vselect <- vars::VARselect(Canada, lag.max = 5, type = 'const')$selection[1]
d.var <- vars::VAR(Canada, p = d.vselect, type = 'const')
ggplot2::fortify(stats::predict(d.var, n.ahead = 50))
}

