% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslib.R
\name{get.dtindex.continuous}
\alias{get.dtindex.continuous}
\title{Get \code{Date} \code{vector} continue to \code{ts} index}
\usage{
get.dtindex.continuous(data, length, is.tsp = FALSE, is.date = NULL)
}
\arguments{
\item{data}{\code{ts} instance}

\item{length}{A number to continue}

\item{is.tsp}{Logical frag whether data is \code{tsp} itself or not}

\item{is.date}{Logical frag indicates whether the \code{stats::ts} is date or not.
If not provided, regard the input as date when the frequency is 4 or 12.}
}
\value{
vector
}
\description{
Get \code{Date} \code{vector} continue to \code{ts} index
}
\examples{
ggfortify:::get.dtindex.continuous(AirPassengers, length = 10)
ggfortify:::get.dtindex.continuous(UKgas, length = 10)
}
