% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslib.R
\name{get.dtindex}
\alias{get.dtindex}
\title{Convert \code{ts} index to \code{Date} \code{vector}}
\usage{
get.dtindex(data, is.tsp = FALSE, is.date = NULL)
}
\arguments{
\item{data}{\code{ts} instance}

\item{is.tsp}{Logical frag whether data is \code{tsp} itself or not}

\item{is.date}{Logical frag indicates whether the \code{stats::ts} is date or not.
If not provided, regard the input as date when the frequency is 4 or 12.}
}
\value{
vector
}
\description{
Convert \code{ts} index to \code{Date} \code{vector}
}
\examples{
ggfortify:::get.dtindex(AirPassengers)
ggfortify:::get.dtindex(UKgas)
ggfortify:::get.dtindex(Nile, is.date = FALSE)
ggfortify:::get.dtindex(Nile, is.date = TRUE)
}
