% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslib.R
\name{ggcpgram}
\alias{ggcpgram}
\title{Plots a cumulative periodogram}
\usage{
ggcpgram(
  ts,
  taper = 0.1,
  colour = "#000000",
  linetype = "solid",
  conf.int = TRUE,
  conf.int.colour = "#0000FF",
  conf.int.linetype = "dashed",
  conf.int.fill = NULL,
  conf.int.alpha = 0.3
)
}
\arguments{
\item{ts}{\code{stats::ts} instance}

\item{taper}{Proportion tapered in forming the periodogram}

\item{colour}{Line colour}

\item{linetype}{Line type}

\item{conf.int}{Logical flag indicating whether to plot confidence intervals}

\item{conf.int.colour}{line colour for confidence intervals}

\item{conf.int.linetype}{line type for confidence intervals}

\item{conf.int.fill}{fill colour for confidence intervals}

\item{conf.int.alpha}{alpha for confidence intervals}
}
\value{
ggplot
}
\description{
Plots a cumulative periodogram
}
\examples{
ggcpgram(AirPassengers)
}
