% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\docType{class}
\name{ggmultiplot-class}
\alias{ggmultiplot-class}
\alias{length,ggmultiplot-method}
\alias{[,ggmultiplot,ANY,ANY,ANY-method}
\alias{[,ggmultiplot-method}
\alias{[[,ggmultiplot-method}
\alias{[<-,ggmultiplot,ANY,ANY,ANY-method}
\alias{[<-,ggmultiplot-method}
\alias{[[<-,ggmultiplot-method}
\title{An S4 class to hold multiple \code{ggplot2::ggplot} instances}
\usage{
\S4method{length}{ggmultiplot}(x)

\S4method{[}{ggmultiplot,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{ggmultiplot}(x, i, j, ..., drop)

\S4method{[}{ggmultiplot,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{[[}{ggmultiplot}(x, i, j, ...) <- value
}
\arguments{
\item{x}{\code{ggmultiplot}}

\item{i}{elements to extract or replace}

\item{j}{not used}

\item{...}{not used}

\item{drop}{not used}

\item{value}{value to be set}
}
\description{
An S4 class to hold multiple \code{ggplot2::ggplot} instances
}
\section{Slots}{

\describe{
\item{\code{plots}}{List of \code{ggplot2::ggplot} instances}

\item{\code{ncol}}{Number of grid columns}

\item{\code{nrow}}{Number of grid rows}
}}

