% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggtsdiag}
\alias{ggtsdiag}
\title{Plots time-series diagnostics.}
\usage{
ggtsdiag(object, gof.lag = 10, conf.int = TRUE,
  conf.int.colour = "#0000FF", conf.int.linetype = "dashed",
  conf.int.fill = NULL, conf.int.alpha = 0.3, ...)
}
\arguments{
\item{object}{A fitted time-series model}

\item{gof.lag}{The maximum number of lags for a Portmanteau goodness-of-fit test}

\item{conf.int}{Logical flag indicating whether to plot confidence intervals}

\item{conf.int.colour}{Line colour for confidence intervals}

\item{conf.int.linetype}{Line type for confidence intervals}

\item{conf.int.fill}{Fill colour for confidence intervals}

\item{conf.int.alpha}{Alpha for confidence intervals}
}
\value{
ggplot
}
\description{
Plots time-series diagnostics.
}
\examples{
library(ggplot2)
ggtsdiag(arima(AirPassengers))
}

