% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslib.R
\name{ggtsdiag}
\alias{ggtsdiag}
\title{Plots time-series diagnostics}
\usage{
ggtsdiag(
  object,
  gof.lag = 10,
  conf.int = TRUE,
  conf.int.colour = "#0000FF",
  conf.int.linetype = "dashed",
  conf.int.fill = NULL,
  conf.int.alpha = 0.3,
  ad.colour = "#888888",
  ad.linetype = "dashed",
  ad.size = 0.2,
  nrow = NULL,
  ncol = 1,
  ...
)
}
\arguments{
\item{object}{A fitted time-series model}

\item{gof.lag}{The maximum number of lags for a Portmanteau goodness-of-fit test}

\item{conf.int}{Logical flag indicating whether to plot confidence intervals}

\item{conf.int.colour}{line colour for confidence intervals}

\item{conf.int.linetype}{line type for confidence intervals}

\item{conf.int.fill}{fill colour for confidence intervals}

\item{conf.int.alpha}{alpha for confidence intervals}

\item{ad.colour}{Line colour for additional lines}

\item{ad.linetype}{Line type for additional lines}

\item{ad.size}{Fill colour for additional lines}

\item{nrow}{Number of facet/subplot rows}

\item{ncol}{Number of facet/subplot columns}

\item{...}{other keywords}
}
\value{
ggplot
}
\description{
Plots time-series diagnostics
}
\examples{
ggtsdiag(arima(AirPassengers))
}
