% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslib.R
\name{is.univariate}
\alias{is.univariate}
\title{Check if Validates number of \code{ts} variates}
\usage{
is.univariate(data, raise = TRUE)
}
\arguments{
\item{data}{\code{ts} instance}

\item{raise}{Logical flag whether raise an error}
}
\value{
logical
}
\description{
Check if Validates number of \code{ts} variates
}
\examples{
ggfortify:::is.univariate(AirPassengers)
}
