% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{parse.formula}
\alias{parse.formula}
\title{Expand \code{stats::formula} expression}
\usage{
parse.formula(formula)
}
\arguments{
\item{formula}{\code{stats::formula} instance}
}
\value{
list
}
\description{
Expand \code{stats::formula} expression
}
\examples{
ggfortify:::parse.formula(y ~ x)
}

