% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{+,ggmultiplot,ANY-method}
\alias{+,ggmultiplot,ANY-method}
\title{Generic add operator for \code{ggmultiplot}}
\usage{
\S4method{+}{ggmultiplot,ANY}(e1, e2)
}
\arguments{
\item{e1}{first argument}

\item{e2}{second argument}
}
\value{
\code{ggmultiplot}
}
\description{
Generic add operator for \code{ggmultiplot}
}
