% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{print,ggmultiplot-method}
\alias{print,ggmultiplot-method}
\title{Generic print function for \code{ggmultiplot}}
\usage{
\S4method{print}{ggmultiplot}(x)
}
\arguments{
\item{x}{\code{ggmultiplot}}
}
\description{
Generic print function for \code{ggmultiplot}
}
