% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslib.R
\name{residuals.ar}
\alias{residuals.ar}
\title{Calculate residuals for \code{stats::ar}}
\usage{
\method{residuals}{ar}(object, ...)
}
\arguments{
\item{object}{\code{stats::ar} instance}

\item{...}{other keywords}
}
\value{
ts Residuals extracted from the object object.
}
\description{
Calculate residuals for \code{stats::ar}
}
\examples{
residuals(ar(WWWusage))
}
