% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{show,ggmultiplot-method}
\alias{show,ggmultiplot-method}
\title{Generic show function for \code{ggmultiplot}}
\usage{
\S4method{show}{ggmultiplot}(object)
}
\arguments{
\item{object}{\code{ggmultiplot}}
}
\description{
Generic show function for \code{ggmultiplot}
}
