% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unscale}
\alias{unscale}
\title{Backtransform \code{scale}-ed object}
\usage{
unscale(data, center = NULL, scale = NULL)
}
\arguments{
\item{data}{Scaled data}

\item{center}{Centered vector}

\item{scale}{Scale vector}
}
\value{
data.frame
}
\description{
Backtransform \code{scale}-ed object
}
\examples{
df <- iris[-5]
ggfortify::unscale(base::scale(df))
}
